DDF-2467 Extract UnconfiguredSolrCatalogProvider inner class
